refactor(create-a-container): split app construction from process startup
server.js previously built the express app inside main() and called
listen(), making the real app unreachable from tests — tests/helpers/
app.js had to rebuild a partial copy of the middleware stack, which
could silently drift from production.

- app.js: new buildApp({ sessionSecrets, rateLimit, accessLog })
  containing the entire HTTP-visible stack (verbatim from server.js);
  also unref()s the session store's expired-session sweeper so the
  interval never keeps the process alive by itself (fixes jest hanging
  on open handles)
- server.js: now only dotenv + DB-backed session secrets + listen()
- tests/helpers/app.js: consumes the real buildApp with a fixed secret,
  rate limiting off (4xx assertions must not burn the limiter budget),
  and access log off

Suite (23 tests) passes against the real app; boot smoke-tested:
/api/v1/health 200, unauthenticated /api/v1/apikeys 401.
