+++ create-a-container/tests/helpers/app.js
@@ -1,31 +1,26 @@
 /**
- * Minimal express app for supertest — mirrors the /api/v1 mount in server.js
- * (query parser, trust proxy, session) without listen(), morgan, rate
- * limiting, swagger, or the SPA fallback.
+ * Test app — the REAL app from app.js (same middleware stack, session store,
+ * routers, and error handling as production), built with test-friendly
+ * options:
  *
- * Sessions use the in-memory store with a fixed secret; DB-backed session
- * secrets and the sequelize store are deployment concerns, not route
- * behavior. Tests normally authenticate with a Bearer API key (which
- * apiAuth accepts and csrfGuard exempts), so most suites never touch the
- * session at all.
+ *   - fixed session secret (production reads secrets from the DB in server.js)
+ *   - rate limiting off (tests assert on 4xx responses; the limiter would
+ *     start rejecting them after 10)
+ *   - access log off (quiet output)
+ *
+ * Tests normally authenticate with a Bearer API key (apiAuth accepts it and
+ * csrfGuard exempts Bearer-only requests), so no session/CSRF choreography
+ * is needed.
  */
 
-const express = require('express');
-const session = require('express-session');
+const { buildApp: buildRealApp } = require('../../app');
 
 function buildApp() {
-  const app = express();
-  app.set('query parser', 'extended');
-  app.set('trust proxy', 1);
-  app.use(
-    session({
-      secret: 'test-secret',
-      resave: false,
-      saveUninitialized: false,
-    }),
-  );
-  app.use('/api/v1', require('../../routers/api/v1'));
-  return app;
+  return buildRealApp({
+    sessionSecrets: ['test-secret'],
+    rateLimit: false,
+    accessLog: false,
+  });
 }
 
 /** Convenience: supertest .set() args for Bearer-key auth. */
